[android] Move Speed Cam preference into Voice Instruction preference menu
