+++ android/app/src/main/res/values/donottranslate.xml
@@ -23,7 +23,7 @@
   <string name="pref_tts_test_voice" translatable="false">TtsTestVoice</string>
   <string name="pref_tts_info" translatable="false">TtsInfo</string>
   <string name="pref_tts_info_link" translatable="false">TtsInfoLink</string>
-  <string name="pref_speed_cameras" translatable="false">SpeedCameras</string>
+  <string name="pref_tts_speed_cameras" translatable="false">SpeedCameras</string>
   <!-- TODO: Move to another domain. -->
   <string name="tts_info_link" translatable="false">https://mapsme.zendesk.com/hc/en-us/articles/208628985-How-can-I-check-TTS-settings-on-my-Android-device-</string>
   <string name="pref_autodownload" translatable="false">AutoDownloadMap</string>
