+++ android/app/src/main/res/xml/prefs_main.xml
@@ -127,16 +127,8 @@
       android:persistent="false"
       android:order="4">
     </Preference>
-    <ListPreference
-      android:key="@string/pref_speed_cameras"
-      android:title="@string/speedcams_alert_title"
-      app:singleLineTitle="false"
-      android:entries="@array/speed_cameras"
-      android:entryValues="@array/speed_cameras_values"
-      android:defaultValue="@string/auto_enum_value"
-      android:order="5"/>
     <PreferenceScreen
-      android:order="6"
+      android:order="5"
       android:title="@string/driving_options_title">
       <intent
         android:targetClass="app.organicmaps.settings.DrivingOptionsActivity"
