+++ data/strings/strings.txt
@@ -22904,6 +22904,144 @@
     zh-Hans = 加载地图
     zh-Hant = 載入地圖
 
+  [pref_tts_speedcams_auto]
+    comment = Speed camera settings menu option - Warn (about speedcams) if risk of exceeding speed limit
+    tags = android
+    en = Warn if there is a risk of exceeding the speed limit
+    af = Waarsku indien daar ’n risiko is dat die grens oorskry word
+    ar = حذرني  إذا هناك احتمالية تجاوز الحد الأقصى للسرعة
+    az = Sürət həddini aşmaq riski varsa, xəbərdarlıq edin
+    be = Папярэджваць калі ёсць рызыка перавышэння дазволенай хуткасці
+    bg = Предупреждава ако има риск от превишаване на ограничението на скоростта
+    ca = Avisa si hi ha risc d'excedir el límit de velocitat
+    cs = Upozorňovat pokud hrozí riziko překročení rychlosti
+    da = Advar hvis der er en risiko for, at hastighedsgrænsen overskrides
+    de = Warnen wenn die Gefahr einer Geschwindigkeitsüberschreitung besteht
+    el = Να ειδοποιούμαι αν υπάρχει κίνδυνος να υπερβώ το όριο ταχύτητας
+    es = Alertar si existe el riesgo de exceder el límite de velocidad
+    es-MX = Avisar si existe el riesgo de exceder el límite de velocidad
+    et = Hoiatused kui on oht ületada piirkiirust
+    eu = Ohartarazten du abiadura muga gainditzeko arriskua badago
+    fa = در صورتی که خطر تجاوز از سرعت مجاز وجود داشت، نسبت هشدار دهد
+    fi = Varoita jos tuntuu siltä, että nopeusrajan ylitys on mahdollista
+    fr = Alerte sonore en cas de risque de dépassement de la limite de vitesse
+    he = הזהר אם יש חשש להפרזה במהירות
+    hu = Figyelmeztetés ha van kockázat a sebesség korlátozás megsértéséről
+    id = Peringatkan saat terdapat risiko melebihi batas kecepatan
+    it = Avvisare in caso di rischio eccesso velocità
+    ja = 速度制限超過の危険性がある場合には、自動速度違反取締装置について警告します
+    ko = 속도 제한 초과 위험이 있을때 경고하기
+    lt = įspėti jei yra greičio viršijimo rizika
+    mr = वेगमर्यादा ओलांडण्याचा धोका असल्यास चेतावणी द्या
+    nb = Advarsel hvis det er fare for å overskride fartsgrensen
+    nl = Waarschuw als er een risico bestaat dat de snelheidslimiet wordt overschreden
+    pl = Ostrzegaj jeśli istnieje ryzyko przekroczenia ograniczenia prędkości
+    pt = Avisar se houver risco de exceder o limite de velocidade
+    pt-BR = Avisar se houver risco de ultrapassar o limite de velocidade
+    ro = Avertizează dacă există riscul depășirii limitei de viteză
+    ru = Предупреждать если есть риск превышения скоростного лимита
+    sk = Upozornenie ak existuje riziko prekročenia rýchlostného limitu
+    sv = Varna om det finns risk att överskrida hastighetsgränsen
+    sw = Onyo iwapo kuna hatari yoyote ya kuzidi ukomo wa mwendo-kasi
+    th = เตือนเกี่ยวกับกล้องตรวจจับความเร็วหากมีความเป็นได้ว่าจะขับเร็วเกิดกำหนด
+    tr = Hız sınırını aşma riski varsa uyar
+    uk = Попереджати якщо є ризик перевищення швидкісного ліміту
+    vi = Cảnh báo nếu có nguy cơ vượt quá giới hạn tốc độ
+    zh-Hans = 如果存在超出速度限制的风险，则对速度摄像头发出警告
+    zh-Hant = 如果存在超出速度限制的風險，則對速度照相機發出警告
+
+  [pref_tts_speedcams_always]
+    comment = Speed camera settings menu option - Always warn (about speedcams)
+    tags = android
+    en = Always warn
+    af = Waarsku altyd
+    ar = حذرني دائما من
+    az = Həmişə xəbərdar edin
+    be = Заўсёды папярэджваць
+    bg = Винаги се продупреждава
+    ca = Sempre avisa
+    cs = Vždy upozorňovat
+    da = Advar altid
+    de = Immer warnen
+    el = Πάντα να ειδοποιούμαι
+    es = Alertar siempre
+    es-MX = Siempre dar aviso
+    et = Hoiata alati eest
+    eu = Beti abisatu
+    fa = همیشه هشدار دهد
+    fi = Varoita aina
+    fr = Toujours avertir
+    he = הזהר  תמיד
+    hu = Mindig figyelmeztet
+    id = Selalu peringatkan
+    it = Avvisare sempre
+    ja = 自動速度違反取締装置について常に警告します
+    ko = 항상 경고하기
+    lt = Visada įspėti
+    mr = नेहमी चेतावणी द्या
+    nb = Advar alltid
+    nl = Waarschuw altijd
+    pl = Zawsze ostrzegaj
+    pt = Sempre avisar
+    pt-BR = Sempre avisar
+    ro = Avertizează mereu
+    ru = Всегда предупреждать
+    sk = Vždy upozorniť
+    sv = Varna alltid
+    sw = Daima onya
+    th = เตือนเกี่ยวกับกล้องตรวจจับความเร็วเสมอ
+    tr = Her zaman uyar
+    uk = Завжди попереджати
+    vi = Luôn cảnh báo
+    zh-Hans = 始终提醒摄像头
+    zh-Hant = 始終提醒照相機
+
+  [pref_tts_speedcams_never]
+    comment = Speed camera settings menu option - Never warn (about speedcams)
+    tags = android
+    en = Never warn
+    af = Waarsku nooit
+    ar = لا تحذرني من أبداً
+    az = Heç xəbərdar etməyin
+    be = Ніколі не папярэджваць
+    bg = Никога не се предупреждава
+    ca = No n'avisis mai
+    cs = Nikdy neupozorňovat
+    da = Advar aldrig
+    de = Niemals warnen
+    el = Ποτέ να μην ειδοποιούμαι
+    es = No alertar nunca
+    es-MX = Nunca dar aviso
+    et = Ära kunagi hoiata eest
+    eu = Inoiz ez abisatu
+    fa = هرگز  هشدار ندهد
+    fi = Älä koskaan varoita
+    fr = Jamais avertir
+    he = אל תזהיר אף פעם
+    hu = Soha ne figyelmezet
+    id = Jangan pernah peringatkan
+    it = Non avvisare mai
+    ja = 自動速度違反取締装置を警告しません
+    ko = 경고하지 않기
+    lt = Neįspėti
+    mr = कधीही चेतावणी देऊ नका
+    nb = Advar aldri
+    nl = Waarschuw nooit
+    pl = Nigdy nie ostrzegaj
+    pt = Nunca avisar
+    pt-BR = Nunca avisar
+    ro = Nu avertiza niciodată
+    ru = Никогда не предупреждать
+    sk = Nikdy neupozorňovať
+    sv = Varna aldrig
+    sw = Kamwe usionye
+    th = ไม่เคยเตือนเกี่ยวกับกล้องตรวจจับความเร็ว
+    tr = Hiç bir zaman uyarma
+    uk = Ніколи не попереджати
+    vi = Không bao giờ cảnh báo
+    zh-Hans = 永远不会对摄像头发出警告
+    zh-Hant = 永遠不會對照相機發出警告
+
   [speedcams_notice_message]
     tags = ios
     en = Auto - Warn about speedcams if there is a risk of exceeding the speed limit\nAlways - Always warn about speedcams\nNever - Never warn about speedcams
